+++ packages/entities/entities-shared/src/index.ts
@@ -30,3 +30,6 @@ export { useAxios, useDeleteUrlBuilder, useErrors, useExternalLinkCreator, useFe
 
 // Types
 export * from './types'
+
+// Constants
+export * from './constants'