+++ packages/analytics/analytics-geo-map/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.49](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.4.48...@kong-ui-public/analytics-geo-map@1.4.49) (2026-01-16)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [1.4.48](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.4.47...@kong-ui-public/analytics-geo-map@1.4.48) (2026-01-15)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [1.4.47](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.4.46...@kong-ui-public/analytics-geo-map@1.4.47) (2026-01-15)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
+## [1.4.46](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.4.45...@kong-ui-public/analytics-geo-map@1.4.46) (2026-01-14)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-geo-map
+
+
+
+
+
 ## [1.4.45](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-geo-map@1.4.44...@kong-ui-public/analytics-geo-map@1.4.45) (2026-01-13)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-geo-map