+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.1.17](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.1.16...@kong-ui-public/analytics-metric-provider@11.1.17) (2026-01-16)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [11.1.16](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.1.15...@kong-ui-public/analytics-metric-provider@11.1.16) (2026-01-15)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [11.1.15](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.1.14...@kong-ui-public/analytics-metric-provider@11.1.15) (2026-01-15)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [11.1.14](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.1.13...@kong-ui-public/analytics-metric-provider@11.1.14) (2026-01-14)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [11.1.13](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.1.12...@kong-ui-public/analytics-metric-provider@11.1.13) (2026-01-13)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider