+++ packages/entities/entities-upstreams-targets/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.12.43](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.12.42...@kong-ui-public/entities-upstreams-targets@3.12.43) (2026-01-16)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
+## [3.12.42](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.12.41...@kong-ui-public/entities-upstreams-targets@3.12.42) (2026-01-15)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
 ## [3.12.41](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.12.40...@kong-ui-public/entities-upstreams-targets@3.12.41) (2026-01-09)
 
 **Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets